+++ tests/auto/threed/qsphere3d/tst_qsphere3d.cpp
@@ -65,61 +65,61 @@ void tst_QSphere3D::create()
 {
     QSphere3D sphere1;
     QCOMPARE(sphere1.center(), QVector3D(0, 0, 0));
-    QCOMPARE(sphere1.radius(), qreal(1.0f));
+    QCOMPARE(sphere1.radius(), 1.0f);
 
     QSphere3D sphere2(QVector3D(1.0, 2.5f, -4.0f), 5.25f);
     QCOMPARE(sphere2.center(), QVector3D(1.0f, 2.5f, -4.0f));
-    QCOMPARE(sphere2.radius(), qreal(5.25f));
+    QCOMPARE(sphere2.radius(), 5.25f);
 
     sphere2.setCenter(QVector3D(-1.0, -2.5f, 4.0f));
     sphere2.setRadius(10.0f);
     QCOMPARE(sphere2.center(), QVector3D(-1.0f, -2.5f, 4.0f));
-    QCOMPARE(sphere2.radius(), qreal(10.0f));
+    QCOMPARE(sphere2.radius(), 10.0f);
 }
 
 void tst_QSphere3D::contains_data()
 {
     QTest::addColumn<QVector3D>("center");
-    QTest::addColumn<qreal>("radius");
+    QTest::addColumn<float>("radius");
     QTest::addColumn<QVector3D>("point");
     QTest::addColumn<bool>("contained");
 
     QTest::newRow("origin-inside")
-        << QVector3D(0, 0, 0) << qreal(0)
+        << QVector3D(0, 0, 0) << 0.0f
         << QVector3D(0, 0, 0) << true;
     QTest::newRow("origin-outside")
-        << QVector3D(0, 0, 0) << qreal(0)
+        << QVector3D(0, 0, 0) << 0.0f
         << QVector3D(0, 0.25f, 0) << false;
 
     QTest::newRow("general-inside")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(1, 2, 3) << true;
 
     QTest::newRow("general-edge-x")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(6, 2, 3) << true;
     QTest::newRow("general-edge-y")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(1, -3, 3) << true;
     QTest::newRow("general-edge-z")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(1, 2, 8) << true;
 
     QTest::newRow("general-outside-x")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(6.25f, 2, 3) << false;
     QTest::newRow("general-outside-y")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(1, -3.5f, 3) << false;
     QTest::newRow("general-outside-z")
-        << QVector3D(1, 2, 3) << qreal(5)
+        << QVector3D(1, 2, 3) << 5.0f
         << QVector3D(1, 2, 8.25f) << false;
 }
 
 void tst_QSphere3D::contains()
 {
     QFETCH(QVector3D, center);
-    QFETCH(qreal, radius);
+    QFETCH(float, radius);
     QFETCH(QVector3D, point);
     QFETCH(bool, contained);
 
@@ -167,7 +167,7 @@ void tst_QSphere3D::dataStream()
 void tst_QSphere3D::transform()
 {
     QVector3D center(1, 2, 3);
-    qreal radius = 5.25f;
+    float radius = 5.25f;
 
     QMatrix4x4 m;
     m.translate(-1.0f, 2.5f, 5.0f);
@@ -184,7 +184,7 @@ void tst_QSphere3D::transform()
     QCOMPARE(sphere1.center(), sphere3.center());
     QCOMPARE(sphere1.radius(), sphere3.radius());
 
-    qreal tradius = m.mapVector(QVector3D(0, radius, 0)).length();
+    float tradius = m.mapVector(QVector3D(0, radius, 0)).length();
 
     QVERIFY(qFuzzyCompare(sphere1.center(), m * center));
     QCOMPARE(float(sphere1.radius()), float(tradius));
@@ -215,14 +215,14 @@ void tst_QSphere3D::properties()
 
     QSphere3D s = qvariant_cast<QSphere3D>(obj.property("sphere"));
     QCOMPARE(s.center(), QVector3D(1, 2, 3));
-    QCOMPARE(s.radius(), qreal(4.0f));
+    QCOMPARE(s.radius(), 4.0f);
 
     obj.setProperty("sphere",
                     qVariantFromValue(QSphere3D(QVector3D(-1, -2, -3), -4)));
 
     s = qvariant_cast<QSphere3D>(obj.property("sphere"));
     QCOMPARE(s.center(), QVector3D(-1, -2, -3));
-    QCOMPARE(s.radius(), qreal(-4.0f));
+    QCOMPARE(s.radius(), -4.0f);
 }
 
 void tst_QSphere3D::metaTypes()
